feat: remove select-model
